refactor(impact): move KiloClaw referral module
