+++ apps/web/src/app/admin/api/users/[id]/kiloclaw-referral-eligibility/route.test.ts
@@ -5,7 +5,7 @@ import {
   type User,
 } from '@kilocode/db';
 
-import { processPersonalKiloClawPaidConversion } from '@/lib/kiloclaw-referrals';
+import { processPersonalKiloClawPaidConversion } from '@/lib/impact/kiloclaw-referrals';
 import { resolveCurrentPersonalSubscriptionRow } from '@/lib/kiloclaw/current-personal-subscription';
 import { getUserFromAuth } from '@/lib/user.server';
 
@@ -32,7 +32,7 @@ jest.mock('@/lib/kiloclaw/current-personal-subscription', () => ({
   resolveCurrentPersonalSubscriptionRow: jest.fn(),
 }));
 
-jest.mock('@/lib/kiloclaw-referrals', () => ({
+jest.mock('@/lib/impact/kiloclaw-referrals', () => ({
   processPersonalKiloClawPaidConversion: jest.fn(),
 }));
