+++ apps/web/src/app/admin/api/users/[id]/kiloclaw-referral-eligibility/route.ts
@@ -4,7 +4,7 @@ import { insertKiloClawSubscriptionChangeLog } from '@kilocode/db';
 
 import { db } from '@/lib/drizzle';
 import { resolveCurrentPersonalSubscriptionRow } from '@/lib/kiloclaw/current-personal-subscription';
-import { processPersonalKiloClawPaidConversion } from '@/lib/kiloclaw-referrals';
+import { processPersonalKiloClawPaidConversion } from '@/lib/impact/kiloclaw-referrals';
 import { getUserFromAuth } from '@/lib/user.server';
 
 const OverrideBodySchema = z.object({
