+++ apps/web/src/app/api/internal/kiloclaw/billing-side-effects/route.test.ts
@@ -2,7 +2,7 @@ import { NextRequest } from 'next/server';
 import { send as sendEmail } from '@/lib/email';
 import { maybePerformAutoTopUp } from '@/lib/autoTopUp';
 import { enqueueAffiliateEventForUser } from '@/lib/impact/affiliate-events';
-import { processPersonalKiloClawPaidConversion } from '@/lib/kiloclaw-referrals';
+import { processPersonalKiloClawPaidConversion } from '@/lib/impact/kiloclaw-referrals';
 
 jest.mock('@/lib/config.server', () => ({
   INTERNAL_API_SECRET: 'internal-secret',
@@ -36,7 +36,7 @@ jest.mock('@/lib/impact/affiliate-events', () => ({
   enqueueAffiliateEventForUser: jest.fn(),
 }));
 
-jest.mock('@/lib/kiloclaw-referrals', () => ({
+jest.mock('@/lib/impact/kiloclaw-referrals', () => ({
   processPersonalKiloClawPaidConversion: jest.fn(),
 }));
