+++ apps/web/src/lib/stripe.ts
@@ -56,7 +56,7 @@ import {
   handleKiloClawInvoicePaid,
 } from '@/lib/kiloclaw/stripe-handlers';
 import { enqueueImpactSaleReversalForCharge } from '@/lib/impact/affiliate-events';
-import { markPersonalKiloClawReferralPaymentAdverse } from '@/lib/kiloclaw-referrals';
+import { markPersonalKiloClawReferralPaymentAdverse } from '@/lib/impact/kiloclaw-referrals';
 import { invoiceLooksLikeKiloClawByPriceId } from '@/lib/kiloclaw/stripe-invoice-classifier.server';
 import { reportEvents } from '@/lib/ai-gateway/abuse-service';
 import {
